requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
psycopg2-binary>=2.9.0
openpyxl>=3.1.0
